dating-bot-render/requirements.txt
aiogram==2.25.1
Flask==2.3.3
waitress==2.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pytz==2023.3.post1
requests==2.31.0
geopy==2.4.1
aiohttp==3.9.0